requirements.txt
torch
torchvision
requests
pandas
tqdm
numpy
pycocoevalcap
decord
pillow
einops
easydict
timm
opencv-python
flow_vis
h5py
av
tensorboard
terminaltables
fvcore
lz4
ftfy
regex
colorlog
tabulate
joblib>=1.3.0 # require joblib>=1.3.0 to return a generator
pyyaml